Continue to move utils to common
